+++ src/integration-test/java/de/communicode/communikey/CommunikeyIntegrationTest.java
@@ -15,6 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
+
 package de.communicode.communikey;
 
 /**
