+++ src/integration-test/java/de/communicode/communikey/api/ApiRootIt.java
@@ -15,6 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
+
 package de.communicode.communikey.api;
 
 import static de.communicode.communikey.CommunikeyApplication.COMMUNIKEY_REST_API_VERSION;
@@ -33,53 +34,53 @@ import org.springframework.http.HttpStatus;
  * @author sgreb@communicode.de
  * @since 0.4.0
  */
-public class ApiRootIT extends IntegrationBaseTest {
+public class ApiRootIt extends IntegrationBaseTest {
 
     @Test
     public void testGetVersionAsAdmin() {
         given()
-                .auth().oauth2(adminUserOAuth2AccessToken)
-                .param(RequestParameter.API_VERSION)
+            .auth().oauth2(adminUserOAuth2AccessToken)
+            .param(RequestParameter.API_VERSION)
         .when()
-                .get(RequestMappings.API)
+            .get(RequestMappings.API)
         .then()
-                .statusCode(HttpStatus.OK.value())
-                .body("version", equalTo(COMMUNIKEY_REST_API_VERSION));
+            .statusCode(HttpStatus.OK.value())
+            .body("version", equalTo(COMMUNIKEY_REST_API_VERSION));
     }
 
     @Test
     public void testGetVersionAsUser() {
         given()
-                .auth().oauth2(userOAuth2AccessToken)
-                .param(RequestParameter.API_VERSION)
+            .auth().oauth2(userOAuth2AccessToken)
+            .param(RequestParameter.API_VERSION)
         .when()
-                .get(RequestMappings.API)
+            .get(RequestMappings.API)
         .then()
-                .statusCode(HttpStatus.OK.value())
-                .body("version", equalTo(COMMUNIKEY_REST_API_VERSION));
+            .statusCode(HttpStatus.OK.value())
+            .body("version", equalTo(COMMUNIKEY_REST_API_VERSION));
     }
 
     @Test
     public void testIsPrivilegedAsAdmin() {
         given()
-                .auth().oauth2(adminUserOAuth2AccessToken)
-                .param(RequestParameter.API_ME)
+            .auth().oauth2(adminUserOAuth2AccessToken)
+            .param(RequestParameter.API_ME)
         .when()
-                .get(RequestMappings.API)
+            .get(RequestMappings.API)
         .then()
-                .statusCode(HttpStatus.OK.value())
-                .body("privileged", equalTo(true));
+            .statusCode(HttpStatus.OK.value())
+            .body("privileged", equalTo(true));
     }
 
     @Test
     public void testIsPrivilegedAsUser() {
         given()
-                .auth().oauth2(userOAuth2AccessToken)
-                .param(RequestParameter.API_ME)
+            .auth().oauth2(userOAuth2AccessToken)
+            .param(RequestParameter.API_ME)
         .when()
-                .get(RequestMappings.API)
+            .get(RequestMappings.API)
         .then()
-                .statusCode(HttpStatus.OK.value())
-                .body("privileged", equalTo(false));
+            .statusCode(HttpStatus.OK.value())
+            .body("privileged", equalTo(false));
     }
 }
